spec: fold xml_containment_spec.lua into xml_spec.lua
spec/wowless/modules/*_spec.lua files correspond 1:1 with
wowless/modules/*.lua -- there's no wowless/modules/xml_containment.lua,
just wowless/modules/xml.lua, so both specs exercising it belong in the
one file. Folds the exhaustive per-tag containment check in as a nested
describe('every tag as a child of Ui', ...) alongside the existing
hand-picked cases, reusing the same per-product datalua/parse rather
than constructing a second parser instance.
